oscore_cbor.c: Fix handling large negative integers
Exclude OSS-FUZZ seed corpus files from pre-commit hooks.

+++ include/oscore/oscore_cbor.h
@@ -105,7 +105,7 @@ size_t oscore_cbor_put_negative(uint8_t **buffer, size_t *buf_size, int64_t valu
 
 uint8_t oscore_cbor_get_next_element(const uint8_t **buffer, size_t *buf_size);
 
-size_t oscore_cbor_get_element_size(const uint8_t **buffer, size_t *buf_size);
+uint64_t oscore_cbor_get_element_size(const uint8_t **buffer, size_t *buf_size);
 
 uint8_t oscore_cbor_elem_contained(const uint8_t *data, size_t *buf_size,
                                    uint8_t *end);
